Fix: show Edit/Delete buttons to author or admin and secure permissions

+++ blog/views.py
@@ -1,12 +1,15 @@
 from django.shortcuts import render, redirect, get_object_or_404
 from django.contrib.auth.decorators import login_required
 from django.contrib import messages
+from django.http import HttpResponseForbidden
 from .models import BlogPost
 from .forms import BlogPostForm
 
+
 def blog_list(request):
     posts = BlogPost.objects.all().order_by('-created_at')
-    return render(request, 'blog/blog_list.html', {'posts': posts})
+    return render(request, 'blog/blog_list.html', {'posts': posts, 'user': request.user})
+
 
 @login_required
 def create_post(request):
@@ -24,9 +27,12 @@ def create_post(request):
         form = BlogPostForm()
     return render(request, 'blog/blog_form.html', {'form': form})
 
+
 @login_required
 def edit_post(request, pk):
     post = get_object_or_404(BlogPost, pk=pk)
+    if request.user != post.author and not request.user.is_superuser:
+        return HttpResponseForbidden("You are not allowed to edit this post.")
     if request.method == 'POST':
         form = BlogPostForm(request.POST, request.FILES, instance=post)
         if form.is_valid():
@@ -37,20 +43,21 @@ def edit_post(request, pk):
             messages.error(request, 'Please correct the errors below.')
     else:
         form = BlogPostForm(instance=post)
-    return render(request, 'blog/blog_form.html', {'form': form})
+    return render(request, 'blog/blog_form.html', {'form': form, 'post': post})
+
 
 @login_required
 def delete_post(request, pk):
     post = get_object_or_404(BlogPost, pk=pk)
+    if request.user != post.author and not request.user.is_superuser:
+        return HttpResponseForbidden("You are not allowed to delete this post.")
     if request.method == 'POST':
         post.delete()
         messages.success(request, 'Post deleted successfully!')
         return redirect('blog:blog_list')
     return render(request, 'blog/blog_confirm_delete.html', {'post': post})
 
-from django.shortcuts import get_object_or_404, render
-from .models import BlogPost
 
 def blog_detail(request, pk):
     post = get_object_or_404(BlogPost, pk=pk)
-    return render(request, 'blog/blog_detail.html', {'post': post})
+    return render(request, 'blog/blog_detail.html', {'post': post, 'user': request.user})
